Moved the different exported classes into the correct namespace
